disable debug options from being passed to child
When you run an application that uses sails-db-migrate under debug, the old behavior would pass the --debug-brk=<PORT> command line option to the child, effectively making both parent and child try to debug on the same port. This behavior was causing interactive debugging to die when the second process started. This change prevents the --debug-brk option from being sent to the child, preventing this problem.